#11 [FEAT] AttendanceCodeApiController 코드를 통한 출결 등록 추가

+++ src/main/java/com/jaramgroupware/mms/domain/attendance/AttendanceRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttendanceRepository extends JpaRepository<Attendance,Integer> {
+public interface AttendanceRepository extends JpaRepository<Attendance,Long> {
 }
